64-bit, 10.5
git-svn-id: https://svn.code.sf.net/p/osirix/code/osirix@2803 4e9e6e3d-4551-47eb-9dbe-a6127f9400a1

+++ DCM Framework/DCMRecord.m
@@ -235,8 +235,9 @@
 }
 
 - (void)parseRecordSequence:(NSArray *)recordSequence recordType:(NSString *)recordType{
-	int i = 0;
-	int index, nextIndex;
+	NSInteger i = 0;
+	NSInteger index, nextIndex;
+	
 	NSMutableIndexSet *indexSet = [NSMutableIndexSet indexSet];
 	for (i = 0; i < [recordSequence count]; i++) {
 		DCMObject *record = [recordSequence objectAtIndex:i];
